Rename "text-input" to "legacy-text-input"
