+++ catalog/input/typeahead.md
@@ -14,8 +14,8 @@ Source: [`react-bootstrap-typeahead`](https://github.com/Faithlife/react-bootstr
 [More component demos](http://ericgio.github.io/react-bootstrap-typeahead/)
 
 ```
-import { Typeahead } from '@faithlife/styled-ui/dist/text-input';
-import '@faithlife/styled-ui/dist/text-input.css';
+import { Typeahead } from '@faithlife/styled-ui/dist/legacy-text-input';
+import '@faithlife/styled-ui/dist/legacy-text-input.css';
 ```
 
 ```react
