+++ webpack.config.js
@@ -10,7 +10,7 @@ module.exports = {
 		main: './index.js',
 		v6: './index-v6.js',
 		'ag-grid': './components/grid/index.js',
-		'text-input': './components/text-input/index.js',
+		'legacy-text-input': './components/legacy-text-input/index.js',
 		'text-input-v2': './components/text-input-v2/index.js',
 		'group-selector': './components/group-selector/index.js',
 		'share-dialog': './components/share-dialog/index.js',
